chore: move and rename unit tests for orgs-v5
