+++ packages/orgs-v5/package.json
@@ -72,6 +72,7 @@
     "postpublish": "rm oclif.manifest.json",
     "posttest": "yarn lint",
     "prepack": "oclif manifest",
+    "test": "nyc mocha ./test/unit/**/*.unit.test.js",
     "test:acceptance": "mocha './test/acceptance/**/*.acceptance.test.js'",
     "version": "oclif readme && git add README.md"
   }
