+++ packages/orgs-v5/test/unit/commands/access/add.unit.test.js
@@ -1,10 +1,10 @@
 'use strict'
 /* globals beforeEach afterEach context cli nock expect */
 
-let cmd = require('../../../commands/access/add')[0]
-let error = require('../../../lib/error')
-let assertExit = require('../../assert_exit')
-let unwrap = require('../../unwrap')
+let cmd = require('../../../../commands/access/add')[0]
+let error = require('../../../../lib/error')
+let assertExit = require('../../../assert_exit')
+let unwrap = require('../../../unwrap')
 let stubGet = require('../../stub/get')
 let stubPost = require('../../stub/post')
 let apiGet
