+++ packages/orgs-v5/test/unit/commands/access/update.unit.test.js
@@ -1,10 +1,10 @@
 'use strict'
 /* globals beforeEach afterEach context cli nock expect */
 
-let cmd = require('../../../commands/access/update')
-let error = require('../../../lib/error')
-let assertExit = require('../../assert_exit')
-let unwrap = require('../../unwrap')
+let cmd = require('../../../../commands/access/update')
+let error = require('../../../../lib/error')
+let assertExit = require('../../../assert_exit')
+let unwrap = require('../../../unwrap')
 let stubPatch = require('../../stub/patch')
 let stubGet = require('../../stub/get')
 let apiPatchAppCollaborators
