+++ packages/orgs-v5/test/unit/commands/orgs/open.unit.test.js
@@ -12,7 +12,7 @@ describe('heroku org:open', () => {
   beforeEach(() => {
     apiGetOrgInfo = stubGet.teamInfo()
     openStub = sinon.stub(cli, 'open').callsFake(() => {})
-    cmd = proxyquire('../../../commands/orgs/open', {
+    cmd = proxyquire('../../../../commands/orgs/open', {
       cli: openStub,
     })
     cli.mockConsole()
